requirements.txt
appdirs==1.4.4
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
cymem==2.0.5
Flask==2.0.1
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.2
packaging==21.0
#pathy==0.6.0
pathy
#pkg_resources==0.0.0
#pkg_resources
preshed==3.0.5
pydantic==1.8.2
PyHyphen==4.0.1
pyparsing==2.4.7
python-docx==0.8.11
regex==2021.8.28
requests==2.26.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
textblob==0.15.3
textblob-de==0.4.3
thinc==8.0.10
tqdm==4.62.2
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
WTForms==2.3.3
